+++ tests/Application/Query/Navigation/Submenu/UserCtaSubmenuViewQueryHandlerTest.php
@@ -8,30 +8,39 @@ use Proximum\Vimeet\Application\Query\Navigation\Submenu\UserCtaSubmenuViewQuery
 use Proximum\Vimeet\Application\Query\Navigation\Submenu\UserCtaSubmenuViewQueryHandler;
 use Proximum\Vimeet\Application\View\Navigation\SubmenuButtonView;
 use Proximum\Vimeet\Domain\Event\ExtraParameter\Type as ExtraParameterType;
+use Proximum\Vimeet\Domain\User\Event\ExtraData\Type as TypeExtraData;
 use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Domain\Model\Participant;
 use Proximum\Vimeet\Domain\Model\Sheet;
 use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Model\User;
 use Proximum\Vimeet\Domain\Repository\Event\ExtraParameterRepositoryInterface;
+use Proximum\Vimeet\Domain\Repository\User\Event\ExtraDataRepositoryInterface;
 
 class UserCtaSubmenuViewQueryHandlerTest extends TestCase
 {
     public function testCustomButton(): void
     {
-        $type = $this->prophesize(Type::class);
-        $type->getId()->shouldBeCalled()->willReturn(746);
-        $sheet = $this->prophesize(Sheet::class);
-        $sheet->getType()->shouldBeCalled()->willReturn($type->reveal());
+        $participant = $this->prophesize(Participant::class);
+        $participant->getId()->shouldBeCalled()->willReturn(123);
         $user = $this->prophesize(User::class);
         $user->getId()->willReturn(42);
         $user->getEmail()->willReturn('test@yahoo.fr');
         $event = $this->prophesize(Event::class);
+        $type = $this->prophesize(Type::class);
+        $type->getId()->shouldBeCalled()->willReturn(746);
+        $sheet = $this->prophesize(Sheet::class);
+        $sheet->getId()->shouldBeCalled()->willReturn(123456);
+        $sheet->getType()->shouldBeCalled()->willReturn($type->reveal());
+        $sheet->getUserParticipant($user->reveal())->willReturn($participant->reveal());
 
         $extraParameterRepository = $this->prophesize(ExtraParameterRepositoryInterface::class);
+        $extraData = $this->prophesize(User\Event\ExtraData::class);
+        $extraData->getValue()->shouldBeCalled()->willReturn(123456789);
         $extraParameter = $this->prophesize(Event\ExtraParameter::class);
         $extraParameter->getValue()->shouldBeCalled()->willReturn('
         {
-             "link": "https://example.net/%userId%/%userEmail%",
+             "link": "https://example.net/%userId%/%userEmail%/%participantId%/%techEventIdContact%/%sheetId%",
              "concerned_type_ids": [689, 746, 747, 748, 749, 750],
              "button-label": {
              "fr": "Mon bouton",
@@ -40,8 +49,12 @@ class UserCtaSubmenuViewQueryHandlerTest extends TestCase
         }');
         $extraParameterRepository->findByEventAndType($event->reveal(), ExtraParameterType::TYPE_CUSTOM_BUTTON)->shouldBeCalled()->willReturn($extraParameter->reveal());
 
+        $extraDataRepository = $this->prophesize(ExtraDataRepositoryInterface::class);
+        $extraDataRepository->getExtraDataForEventNameAndUser($event->reveal(),TypeExtraData::IMPORTED_FROM_TECH_EVENT, $user->reveal())->shouldBeCalled()->willReturn($extraData->reveal());
+
         $userCtaSubmenuViewQueryHandler = new UserCtaSubmenuViewQueryHandler(
-            $extraParameterRepository->reveal()
+            $extraParameterRepository->reveal(),
+            $extraDataRepository->reveal()
         );
 
         $result = $userCtaSubmenuViewQueryHandler->handle(
@@ -56,17 +69,72 @@ class UserCtaSubmenuViewQueryHandlerTest extends TestCase
         $expectedCustomButtonSubmenuButtonView = new SubmenuButtonView(
         Category::CUSTOM_BUTTON_ICON,
             'Mon bouton',
-            'https://example.net/42/test%40yahoo.fr',
+            'https://example.net/42/test%40yahoo.fr/123/25f9e794323b453885f5181f1b624d0b/123456',
             false,
             null,
             false,
             ['target' => '_blank']
     );
 
-        $this->assertEquals($expectedCustomButtonSubmenuButtonView, $result);
+        self::assertEquals($expectedCustomButtonSubmenuButtonView, $result);
     }
 
-    public function testCustomButtonNull(): void
+    public function testCustomButtonWhenParticipantIdNotNeeded(): void
+    {
+        $user = $this->prophesize(User::class);
+        $user->getId()->willReturn(42);
+        $user->getEmail()->willReturn('test@yahoo.fr');
+        $event = $this->prophesize(Event::class);
+        $type = $this->prophesize(Type::class);
+        $type->getId()->shouldBeCalled()->willReturn(746);
+        $sheet = $this->prophesize(Sheet::class);
+        $sheet->getId()->willReturn(123456);
+        $sheet->getType()->shouldBeCalled()->willReturn($type->reveal());
+        $sheet->getUserParticipant($user->reveal())->willReturn(null);
+
+        $extraParameterRepository = $this->prophesize(ExtraParameterRepositoryInterface::class);
+        $extraParameter = $this->prophesize(Event\ExtraParameter::class);
+        $extraParameter->getValue()->shouldBeCalled()->willReturn('
+        {
+             "link": "https://example.net/%userId%/%userEmail%",
+             "concerned_type_ids": [689, 746, 747, 748, 749, 750],
+             "button-label": {
+             "fr": "Mon bouton",
+             "en": "My button"
+             }
+        }');
+        $extraParameterRepository->findByEventAndType($event->reveal(), ExtraParameterType::TYPE_CUSTOM_BUTTON)->shouldBeCalled()->willReturn($extraParameter->reveal());
+
+        $extraDataRepository = $this->prophesize(ExtraDataRepositoryInterface::class);
+
+        $userCtaSubmenuViewQueryHandler = new UserCtaSubmenuViewQueryHandler(
+            $extraParameterRepository->reveal(),
+            $extraDataRepository->reveal()
+        );
+
+        $result = $userCtaSubmenuViewQueryHandler->handle(
+            new UserCtaSubmenuViewQuery(
+                $user->reveal(),
+                $event->reveal(),
+                'fr',
+                $sheet->reveal()
+            )
+        );
+
+        $expectedCustomButtonSubmenuButtonView = new SubmenuButtonView(
+            Category::CUSTOM_BUTTON_ICON,
+            'Mon bouton',
+            'https://example.net/42/test%40yahoo.fr',
+            false,
+            null,
+            false,
+            ['target' => '_blank']
+        );
+
+        self::assertEquals($expectedCustomButtonSubmenuButtonView, $result);
+    }
+
+    public function testCustomButtonNullWhenTypeNotConcerned(): void
     {
         $type = $this->prophesize(Type::class);
         $type->getId()->shouldBeCalled()->willReturn(701);
@@ -89,8 +157,11 @@ class UserCtaSubmenuViewQueryHandlerTest extends TestCase
         }');
         $extraParameterRepository->findByEventAndType($event->reveal(), ExtraParameterType::TYPE_CUSTOM_BUTTON)->shouldBeCalled()->willReturn($extraParameter->reveal());
 
+        $extraDataRepository = $this->prophesize(ExtraDataRepositoryInterface::class);
+
         $userCtaSubmenuViewQueryHandler = new UserCtaSubmenuViewQueryHandler(
-            $extraParameterRepository->reveal()
+            $extraParameterRepository->reveal(),
+            $extraDataRepository->reveal()
         );
 
         $result = $userCtaSubmenuViewQueryHandler->handle(
@@ -102,6 +173,78 @@ class UserCtaSubmenuViewQueryHandlerTest extends TestCase
             )
         );
 
-        $this->assertEquals(null, $result);
+        self::assertEquals(null, $result);
+    }
+
+    public function testCustomButtonNullWhenNotConfigured(): void
+    {
+        $sheet = $this->prophesize(Sheet::class);
+        $user = $this->prophesize(User::class);
+        $event = $this->prophesize(Event::class);
+
+        $extraParameterRepository = $this->prophesize(ExtraParameterRepositoryInterface::class);
+        $extraParameterRepository->findByEventAndType($event->reveal(), ExtraParameterType::TYPE_CUSTOM_BUTTON)->shouldBeCalled()->willReturn(null);
+
+        $extraDataRepository = $this->prophesize(ExtraDataRepositoryInterface::class);
+
+        $userCtaSubmenuViewQueryHandler = new UserCtaSubmenuViewQueryHandler(
+            $extraParameterRepository->reveal(),
+            $extraDataRepository->reveal()
+        );
+
+        $result = $userCtaSubmenuViewQueryHandler->handle(
+            new UserCtaSubmenuViewQuery(
+                $user->reveal(),
+                $event->reveal(),
+                'fr',
+                $sheet->reveal()
+            )
+        );
+
+        self::assertNull($result);
+    }
+
+    public function testCustomButtonNullWhenNotParticipant(): void
+    {
+        $user = $this->prophesize(User::class);
+        $user->getId()->willReturn(42);
+        $user->getEmail()->willReturn('test@yahoo.fr');
+        $event = $this->prophesize(Event::class);
+        $type = $this->prophesize(Type::class);
+        $type->getId()->shouldBeCalled()->willReturn(746);
+        $sheet = $this->prophesize(Sheet::class);
+        $sheet->getType()->shouldBeCalled()->willReturn($type->reveal());
+        $sheet->getUserParticipant($user->reveal())->willReturn(null);
+
+        $extraParameterRepository = $this->prophesize(ExtraParameterRepositoryInterface::class);
+        $extraParameter = $this->prophesize(Event\ExtraParameter::class);
+        $extraParameter->getValue()->shouldBeCalled()->willReturn('
+        {
+             "link": "https://example.net/%userId%/%userEmail%/%participantId%/%techEventIdContact%",
+             "concerned_type_ids": [689, 746, 747, 748, 749, 750],
+             "button-label": {
+             "fr": "Mon bouton",
+             "en": "My button"
+             }
+        }');
+        $extraParameterRepository->findByEventAndType($event->reveal(), ExtraParameterType::TYPE_CUSTOM_BUTTON)->shouldBeCalled()->willReturn($extraParameter->reveal());
+
+        $extraDataRepository = $this->prophesize(ExtraDataRepositoryInterface::class);
+
+        $userCtaSubmenuViewQueryHandler = new UserCtaSubmenuViewQueryHandler(
+            $extraParameterRepository->reveal(),
+            $extraDataRepository->reveal()
+        );
+
+        $result = $userCtaSubmenuViewQueryHandler->handle(
+            new UserCtaSubmenuViewQuery(
+                $user->reveal(),
+                $event->reveal(),
+                'fr',
+                $sheet->reveal()
+            )
+        );
+
+        self::assertNull($result);
     }
 }